requirements.txt
streamlit
pydeck
pandas
numpy
geopandas
pyyaml
scikit-learn
xarray
rasterio
rioxarray
pyproj
shapely